Moves requirements inside CE so it can be use in standard
